requirements.txt
crewai 
crewai_tools
load_dotenv
langchain-huggingface
langchain_community
langchain_google_genai